Multiples of a Number
Given numbers x and n, where n is a power of 2, print out the smallest multiple of n which is greater than or equal to x. Do not use division or modulo operator. Solution to challenge on CodeEval.
INPUT: path to a filename containing a comma separated list of two integers, one list per line.